* mostly fixed the project view flaky thing: key thing was that if the project was already selected in the tree browser, it wouldn't activate its panel!  so we have to select something else, then select the target, and now it works *most* of the time * how many ways are there for breakpoints not to take!!! I think I've found all of them, but who knows!!  now always do after script recompile for sure. * Added Toggle Break Enable method on Prog El's, and fixed disable all to actually disable * tried to get the BRK symbol to update on the css Script but I just can't seem to get it to do so.

+++ src/temt/ta/ProgEl_List.cpp
@@ -114,11 +114,13 @@ bool ProgEl_List::BrowserCollapseAll() {
 
 bool ProgEl_List::RemoveIdx(int idx) {
   ProgEl* pel = FastEl(idx);
-  Program* prog = GET_MY_OWNER(Program);
-  if (prog) {
-    ProgBrkPt_List* brk_pts = prog->GetBrkPts();
-    if(brk_pts) {
-      brk_pts->DeleteBrkPt(pel);
+  if(pel->owner == this) {
+    Program* prog = GET_MY_OWNER(Program);
+    if (prog) {
+      ProgBrkPt_List* brk_pts = prog->GetBrkPts();
+      if(brk_pts) {
+        brk_pts->DeleteBrkPt(pel);
+      }
     }
   }
   return inherited::RemoveIdx(idx);
